Structural overhaul
Moved things that were not really components

Streamlined artist search functionality

Moved styles to one folder

Began to work on a card component for widespread use
