+++ components/Footer.js
@@ -1,4 +1,4 @@
-import classes from "./Footer.module.css";
+import classes from "../styles/Footer.module.css";
 
 const Footer = () => {
   return (
