+++ pages/profile/artists.js
@@ -1,9 +1,10 @@
-import UserArtists from "@/components/UserArtists";
 import { useEffect, useState, useContext } from "react";
+import Link from "next/link";
 import getProfileDataType from "../api/getProfileDataType";
 import { SpotifyContext } from "@/context/spotifyContext";
 import Navbar from "@/components/Navbar";
 import Footer from "@/components/Footer";
+import classes from '../../styles/UserArtists.module.css'
 
 const artists = () => {
   const { spotifyTokenState, updateSpotifyToken } = useContext(SpotifyContext);
@@ -24,10 +25,51 @@ const artists = () => {
       }
     });
   }, []);
+
+  const renderArtists = () => {
+    return artists.map((artist) => {
+      return (
+        <li key={artist.id}>
+          <div
+            className={`${classes.card} card col-xl-3 col-md-4 col-12 d-flex align-items-center justify-content-center`}
+          >
+            <Link className={`${classes.link}`} href={`/artists/${artist.id}`}>
+              {artist.images.length ? (
+                <img
+                  className={`${classes.img} card-img`}
+                  src={artist.images[0].url}
+                  alt={`Image of ${artist.name}`}
+                />
+              ) : (
+                <></>
+              )}
+
+              <h2 className={`${classes.artistName} card-title`}>
+                {artist.name}
+              </h2>
+            </Link>
+          </div>
+        </li>
+      );
+    });
+  };
   return (
     <>
       <Navbar />
-      <UserArtists artists={artists} />
+      <div className={`${classes.content}`}>
+        <Link href={"/profile"}>
+          <button className={`${classes.profileButton}`}>
+            Back to profile
+          </button>
+        </Link>
+        <h1>Top Artists</h1>
+        <ul className={`${classes.list}`}>{renderArtists()}</ul>
+        <Link href={"/profile"}>
+          <button className={`${classes.profileButton}`}>
+            Back to profile
+          </button>
+        </Link>
+      </div>
       <Footer />
     </>
   );
