+++ pages/profile/toptracks.js
@@ -1,5 +1,6 @@
-import TopTracks from "@/components/TopTracks";
-import { useState, useEffect, useContext } from "react";
+import classes from '../../styles/TopTracks.module.css'
+import { useState, useEffect, useContext, useRef } from "react";
+import Link from 'next/link';
 import getProfileDataType from "../api/getProfileDataType";
 import { SpotifyContext } from "@/context/spotifyContext";
 import Navbar from "@/components/Navbar";
@@ -28,11 +29,99 @@ const toptracks = () => {
     });
     setLoading(false);
   }, []);
+  const listRefs = useRef([]);
+  const [isVisible, setIsVisible] = useState([]);
+  useEffect(() => {
+    setIsVisible(Array(tracks.length).fill(false));
+  }, [tracks]);
+  useEffect(() => {
+    const observers = listRefs.current.map((ref, index) => {
+      return new IntersectionObserver(
+        (entries) => {
+          entries.forEach((entry) => {
+            if (entry.isIntersecting) {
+              setIsVisible((prevState) => {
+                const newState = [...prevState];
+                newState[index] = true;
+                return newState;
+              });
+            }
+          });
+        },
+        { threshold: 0.5 }
+      );
+    });
+    observers.forEach((observer, index) => {
+      observer.observe(listRefs.current[index]);
+    });
+    return () => {
+      observers.forEach((observer) => observer.disconnect());
+    };
+  }, [isVisible]);
+
+  const renderTracks = () => {
+    return tracks.map((track, index) => {
+      return (
+        <li
+          ref={(item) => (listRefs.current[index] = item)}
+          key={track.id}
+          className={`${isVisible[index] ? classes.card : classes.invisible}`}
+        >
+          <div className={`${classes.cardBody}`}>
+            <img
+              className={`${classes.img} card-img`}
+              src={track.album.images[0].url}
+              alt=""
+            />
+            <div className={`${classes.cardText}`}>
+              <h2 className={`${classes.trackName} card-title`}>
+                {track.name}
+              </h2>
+              <div className={`${classes.artistDiv}`}>
+                {track.artists.length === 1 ? "Artist: " : "Artists: "}{" "}
+                <ul className={`${classes.artistUl}`}>
+                  {track.artists.map((artist) => {
+                    return (
+                      <Link
+                        className={`${classes.artists}`}
+                        key={artist.id}
+                        href={`/artists/${artist.id}`}
+                      >
+                        <li className={`${classes.artistName}`}>
+                          {artist.name}
+                        </li>
+                      </Link>
+                    );
+                  })}
+                </ul>
+              </div>
+            </div>
+          </div>
+        </li>
+      );
+    });
+  };
   return (
     <>
       <Navbar />
       {loading && <Loading />}
-      {!loading && <TopTracks tracks={tracks} />}
+      {!loading && (
+        <div
+          className={`${classes.content} d-flex align-items-center justify-content-center`}
+        >
+          <h1
+            className={`${classes.header} d-flex align-items-center justify-content-center`}
+          >
+            Top Tracks
+          </h1>
+          <ul className={`${classes.list}`}>{renderTracks()}</ul>
+          <Link href={"/profile"}>
+            <button className={`${classes.profileButton}`}>
+              Back to profile
+            </button>
+          </Link>
+        </div>
+      )}
 
       <Footer />
     </>
