+++ styles/Card.module.css
@@ -1,0 +1,5 @@
+.wholeCard {
+    background-color: black;
+    border: 1px solid white;
+    color: white;
+}
