+++ styles/globalStyles.css
@@ -1,0 +1,8 @@
+html {
+    height: 100%;
+}
+
+body {
+    margin: 0;
+    padding: 0;
+}
